* EDIT: cl_intersect.h moved from Engine to CDB.
Deleted its duplicates and removed unnecessary includes from some files.
