+++ code/engine.vc2008/xrCDB/xrCDB.vcxproj
@@ -296,6 +296,7 @@
     <ClCompile Include="xrXRC.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="cl_intersect.h" />
     <ClInclude Include="Frustum.h" />
     <ClInclude Include="StdAfx.h" />
     <ClInclude Include="xrCDB.h" />
