+++ code/engine.vc2008/xrEngine/Feel_Vision.cpp
@@ -4,7 +4,7 @@
 #include "xr_object.h"
 #include "xr_collide_form.h"
 #include "igame_level.h"
-#include "cl_intersect.h"
+#include "../xrCDB/cl_intersect.h"
 
 namespace Feel {
