+++ code/engine.vc2008/xrEngine/xrEngine.vcxproj
@@ -361,7 +361,6 @@
     <ClInclude Include="Properties.h" />
     <ClInclude Include="pure.h" />
     <ClInclude Include="std_classes.h" />
-    <ClInclude Include="cl_intersect.h" />
     <ClInclude Include="XR_IOConsole.h" />
     <ClInclude Include="xr_ioc_cmd.h" />
     <ClInclude Include="Engine.h" />
