+++ code/engine.vc2008/xrEngine/xr_efflensflare.cpp
@@ -9,7 +9,7 @@
 //	Instead of SkeletonCustom:
 #include "bone.h"
 #include "../Include/xrRender/Kinematics.h"
-#include "cl_intersect.h"
+#include "../xrCDB/cl_intersect.h"
 
 #include "../xrServerEntities/object_broker.h"
