+++ code/engine.vc2008/xrPhysics/PHSimpleCharacter.cpp
@@ -2,7 +2,7 @@
 
 #include "PHDynamicData.h"
 #include "ExtendedGeom.h"
-#include "../xrEngine/cl_intersect.h"
+#include "../xrCDB/cl_intersect.h"
 #include "../xrEngine/xr_object_list.h"
 #include "tri-colliderKNoOPC\__aabb_tri.h"
 #include "PHSimpleCharacter.h"
@@ -1371,7 +1371,7 @@ void CPHSimpleCharacter::InitContact(dContact* c, bool	&do_collide, u16 material
 	bool object = (dGeomGetBody(g1) && dGeomGetBody(g2));
 	b_on_object = b_on_object || object;
 
-	////////////////////////нужно сместить колижен!!
+	////////////////////////РЅСѓР¶РЅРѕ СЃРјРµСЃС‚РёС‚СЊ РєРѕР»РёР¶РµРЅ!!
 	//////////////
 	FootProcess(c, do_collide, bo1);
 	if (!do_collide) return;
