+++ code/engine.vc2008/xrRender/xrRender/DetailManager.cpp
@@ -6,7 +6,7 @@
 #pragma hdrstop
 
 #include "DetailManager.h"
-#include "cl_intersect.h"
+
 
 #ifdef _EDITOR
 #	include "ESceneClassList.h"
