+++ code/engine.vc2008/xrRender/xrRenderPC_R3/light_vis.cpp
@@ -1,6 +1,5 @@
 #include "StdAfx.h"
 #include "../xrRender/light.h"
-#include "../../xrEngine/cl_intersect.h"
 
 const	u32	delay_small_min			= 1;
 const	u32	delay_small_max			= 3;
