+++ code/engine.vc2008/xrSound/xrSound.vcxproj
@@ -234,7 +234,6 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="cl_intersect.h" />
     <ClInclude Include="Sound.h" />
     <ClInclude Include="SoundRender.h" />
     <ClInclude Include="stdafx.h" />
